Move URL out of screen_char_t and in to iTermExternalAttributes.
This required adding a migration path from the now-legacy
screen_char_t that had a URL code to the modern one that has
16 bits of unused space. This was motivated by the desire to
support the protected attribute and to add new attributes
more easily in the future. This also gets us 32-bit URL
codes, which ought to be enough for anyone. I also fixed a
bug where NSURLs were saved in VT100Terminal's restorable
state when OSC 8 is in effect (rare, but breaks decoding).
Fix a crash when VT100Grid fails to import a broken
restorable state dictionary. Update DVR to import/export
external attributes. Fix a bug in iTermURLStore that used up
codes unnecessarily quickly.

+++ sources/VT100TerminalDelegate.h
@@ -441,8 +441,8 @@ typedef NS_ENUM(int, VT100TerminalColorIndex) {
 - (NSString *)terminalValueOfVariableNamed:(NSString *)name;
 
 // Links
-- (void)terminalWillEndLinkWithCode:(unsigned short)code;
-- (void)terminalWillStartLinkWithCode:(unsigned short)code;
+- (void)terminalWillEndLinkWithCode:(unsigned int)code;
+- (void)terminalWillStartLinkWithCode:(unsigned int)code;
 
 // Custom escape sequences
 - (void)terminalCustomEscapeSequenceWithParameters:(NSDictionary<NSString *, NSString *> *)parameters
