+++ sources/iTermURLMark.h
@@ -10,5 +10,5 @@
 
 // Invisible marks used to record where URL links are located so they can be freed.
 @interface iTermURLMark : iTermMark
-@property (nonatomic) unsigned short code;
+@property (nonatomic) unsigned int code;
 @end
